feat: initial support for private trackers
This commit adds support for private trackers, although for now only basic user / password authentication is supported. This change is also reflected in the UI as well as the converter script.

Next feature to be implemented will be dynamic configuration popups (reading the tracker options from the yaml definition like Jackett does), and after that I'll add the ability to send custom headers such as cookies and api keys.

This will cause the definition format to change, as extra properties will be required for this.