fix: cache on all endpoint
When I introduced parallel workers for the "/all" torznab enpoint, I introduced a race condition that caused a failure, as all the indexers were trying to retrieve the cached results at the same time.

I've changed how the search works. The indexers now recover the cached results sequentially. The indexers that don't have cached results make a live query in parallel, limited by the MAX_CONCURRENT_SEARCHES env variable.

This approach seems to work correctly in my tests, but as all the features in this app, it has to be considered experimental for the time being.