feat: add filtering
I added a simple regex filter that triggers in there are one or more words in the search enclosed in double quotes.

If that's the case, the results that don't have the quoted strings will be filtered out.

For example, searching:

Mario "Bros"

will return all the results from searching "Mario Bros" as usual, and will filter out all the ones that don't contain "Bros".

This also opens the door for implementing full regex searches in the future, but I'm not really sure if that would be useful or if it would cause compatibility issues with clients...